Improvements in item layout, implementation of receipt date and implementation of SwipeRefresh to reload API items.

+++ app/src/main/java/com/example/digitalreceipts/api/model/Fields.kt
@@ -2,7 +2,7 @@ package com.example.digitalreceipts.api.model
 
 data class Fields(
     var id: String,
-    var date: Int,
+    var date: Long,
     var value: Int,
     var status: Int,
     var authentication: String,
